+++ src/scraper/commandListPage/processCommandListPage.js
@@ -1,3 +1,4 @@
+import puppeteer from 'puppeteer';
 import { getCommandPageData } from './commandScrape.js'
